move ToyADMOS dataset loader to examples folder
