+++ examples/toyconveyordatasetloader.py
